+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 7.4.0 (2022-11-10)
+
+### Enhancements
+
+* Support using objcopy instead of objdump to extract native symbols (when supported by the current NDK).
+  [#484](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/484)
+
+### Bug Fixes
+
+* Fix FileAlreadyExistsException errors when building ReactNative projects with Hermes
+  [#482](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/482)
+
 ## 7.3.1 (2022-10-05)
 
 * Fixed a bug where using ndkBuild generated empty some mapping files which could not be uploaded